LayoutTests/fast/mediastream/RTCPeerConnection-stable-expected.txt

Tests RTCPeerConnection in stable state.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS pc.signalingState is "stable"
PASS pc.setLocalDescription(sessionDescription, finishIfSucceeded, requestFailed1); did not throw exception.
PASS setLocalDescription failed.
PASS errorReason.name is "InvalidSessionDescriptionError"
PASS pc.localDescription threw exception InvalidStateError (DOM Exception 11): The object is in an invalid state..
PASS pc.remoteDescription threw exception InvalidStateError (DOM Exception 11): The object is in an invalid state..
PASS pc.signalingState is "stable"
PASS pc.setLocalDescription(sessionDescription, finishIfSucceeded, requestFailed2); did not throw exception.
PASS setLocalDescription failed.
PASS errorReason.name is "InvalidSessionDescriptionError"
PASS pc.localDescription threw exception InvalidStateError (DOM Exception 11): The object is in an invalid state..
PASS pc.remoteDescription threw exception InvalidStateError (DOM Exception 11): The object is in an invalid state..
PASS pc.signalingState is "stable"
PASS pc.setRemoteDescription(sessionDescription, finishIfSucceeded, requestFailed3); did not throw exception.
PASS setRemoteDescription failed.
PASS errorReason.name is "InvalidSessionDescriptionError"
PASS pc.localDescription threw exception InvalidStateError (DOM Exception 11): The object is in an invalid state..
PASS pc.remoteDescription threw exception InvalidStateError (DOM Exception 11): The object is in an invalid state..
PASS pc.signalingState is "stable"
PASS pc.setRemoteDescription(sessionDescription, finishIfSucceeded, requestFailed4); did not throw exception.
PASS setRemoteDescription failed.
PASS errorReason.name is "InvalidSessionDescriptionError"
PASS pc.localDescription threw exception InvalidStateError (DOM Exception 11): The object is in an invalid state..
PASS pc.remoteDescription threw exception InvalidStateError (DOM Exception 11): The object is in an invalid state..
PASS pc.signalingState is "stable"
PASS pc.setLocalDescription(sessionDescription, requestSucceeded1, finishIfFailed); did not throw exception.
PASS setLocalDescription succeeded.
PASS pc.localDescription.type is "offer"
PASS pc.localDescription.sdp is "local"
PASS pc.remoteDescription threw exception InvalidStateError (DOM Exception 11): The object is in an invalid state..
PASS pc.signalingState is "have-local-offer"
PASS pc.localDescription threw exception InvalidStateError (DOM Exception 11): The object is in an invalid state..
PASS pc.remoteDescription threw exception InvalidStateError (DOM Exception 11): The object is in an invalid state..
PASS pc.signalingState is "stable"
PASS pc.setRemoteDescription(sessionDescription, requestSucceeded2, finishIfFailed); did not throw exception.
PASS setRemoteDescription succeeded.
PASS pc.localDescription threw exception InvalidStateError (DOM Exception 11): The object is in an invalid state..
PASS pc.remoteDescription.type is "offer"
PASS pc.remoteDescription.sdp is "remote"
PASS pc.signalingState is "have-remote-offer"
PASS successfullyParsed is true

TEST COMPLETE
